requirements.txt
dlt[duckdb]==1.17.1
dbt-duckdb==1.10.0
duckdb==1.3.2
python-dotenv==1.1.1
requests==2.32.5
